automation/requirements.txt
firecloud==0.16.25
google-cloud-secret-manager==0.2.0
oauth2client
